Added intern pool, and structure for cards
Intern Pool is a serializable data structure,
It contains two members, the hashmap of strings to ids,
And a string arena,
both are serialized to a file.
The hashmap only has the valid keys serialized.

TODO: make sure that the StrIds do not change if the size of the intern pool changes